requirements.txt
opencv-python
pillow
transformers
torch
